refacto of the includedin method
